refactor!: This is a cleanup before migrating to HTMX RESTFUL
remove javafx client
remove multiple project setup root
now it's server only
