--- client/pom.xml
@@ -1,67 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <groupId>com.bgsix</groupId>
-    <artifactId>client</artifactId>
-    <version>1.0-SNAPSHOT</version>
-    <properties>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>22</maven.compiler.source>
-        <maven.compiler.target>22</maven.compiler.target>
-        <javafx.version>22.0.2</javafx.version>
-        <maven-jlink-plugin.version>3.2.0</maven-jlink-plugin.version>
-        <maven-compiler-plugin.version>3.13.0</maven-compiler-plugin.version>
-    </properties>
-    <dependencies>
-        <dependency>
-            <groupId>org.openjfx</groupId>
-            <artifactId>javafx-controls</artifactId>
-            <version>${javafx.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.openjfx</groupId>
-            <artifactId>javafx-fxml</artifactId>
-            <version>${javafx.version}</version>
-        </dependency>
-    </dependencies>
-    <build>
-        <plugins>
-
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>${maven-compiler-plugin.version}</version>
-            </plugin>
-
-            <plugin>
-                <groupId>org.openjfx</groupId>
-                <artifactId>javafx-maven-plugin</artifactId>
-                <configuration>
-                    <jlinkImageName>client</jlinkImageName>
-                    <launcher>Client</launcher>
-                    <stripDebug>true</stripDebug>
-                    <noHeaderFiles>true</noHeaderFiles>
-                    <noManPages>true</noManPages>
-                    <jlinkImageName>client</jlinkImageName>
-                    <jlinkZipName>clientzip</jlinkZipName>
-                    <mainClass>com.bgsix/com.bgsix.App</mainClass>
-                </configuration>
-                <version>0.0.8</version>
-            </plugin>
-
-        </plugins>
-    </build>
-
-    <!-- Profiles -->
-    <profiles>
-        <!-- Combined Clean and Jlink Profile -->
-        <profile>
-            <id>clean-and-jlink</id>
-            <build>
-                <defaultGoal>clean javafx:jlink</defaultGoal>
-            </build>
-        </profile>
-    </profiles>
-
-</project>
