--- client/src/main/java/com/bgsix/PrimaryController.java
@@ -1,12 +1,0 @@
-package com.bgsix;
-
-import java.io.IOException;
-import javafx.fxml.FXML;
-
-public class PrimaryController {
-
-    @FXML
-    private void switchToSecondary() throws IOException {
-        App.setRoot("secondary");
-    }
-}
