--- client/src/main/java/com/bgsix/SecondaryController.java
@@ -1,12 +1,0 @@
-package com.bgsix;
-
-import java.io.IOException;
-import javafx.fxml.FXML;
-
-public class SecondaryController {
-
-    @FXML
-    private void switchToPrimary() throws IOException {
-        App.setRoot("primary");
-    }
-}
