--- client/src/main/java/module-info.java
@@ -1,9 +1,0 @@
-module com.bgsix {
-    requires javafx.controls;
-    requires javafx.fxml;
-    requires transitive javafx.graphics;
-
-    opens com.bgsix to javafx.fxml;
-
-    exports com.bgsix;
-}
