--- server/src/main/java/com/bgsix/server/dto/BookResponse.java
@@ -1,19 +1,0 @@
-package com.bgsix.server.dto;
-
-import java.util.List;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class BookResponse {
-	private List<BookDTO> books;
-	private int totalPages;
-	private long totalElements;
-
-}
