--- server/src/main/java/com/bgsix/server/model/Book.java
@@ -1,49 +1,0 @@
-package com.bgsix.server.model;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Book {
-	@Id
-	@Column(name = "BOOK_ID")
-	private Long bookId;
-
-	private String title;
-
-	private String language;
-
-	private String author;
-
-	private Float rating;
-
-	private String isbn;
-
-	@Column(name = "BOOK_FORMAT")
-	private String bookFormat;
-
-	private Integer pages;
-
-	@Column(name = "NUM_RATINGS")
-	private Integer numRatings;
-
-	@Column(name = "LIKED_PERCENT")
-	private Short likedPercent;
-
-	private Double price;
-
-	@Column(name = "COVER_IMG")
-	private String coverImg;
-
-	@Column(name = "PUBLICATION_YEAR")
-	private Integer publicationYear;
-
-	@OneToMany(mappedBy = "book", cascade = CascadeType.ALL)
-	private List<BookGenre> bookGenres = new ArrayList<BookGenre>();
-}
