+++ src/main/resources/application.properties
@@ -1,0 +1,23 @@
+spring.application.name=bookmanagement
+# logging.level.org.springframework=DEBUG
+
+# Hibernate properties
+# spring.jpa.show-sql=true
+# spring.jpa.properties.hibernate.format_sql=true
+# logging.level.org.hibernate.SQL=DEBUG
+# logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
+# logging.level.org.hibernate.orm.jdbc.bind=TRACE
+
+
+# Spring Doc Open API
+springdoc.api-docs.path=/api-docs
+
+# Paths to include
+springdoc.pathsToMatch=/api/**
+
+
+# P6Spy
+# Register P6LogFactory to log JDBC events
+decorator.datasource.p6spy.enable-logging=false
+# Use com.p6spy.engine.spy.appender.MultiLineFormat instead of com.p6spy.engine.spy.appender.SingleLineFormat
+decorator.datasource.p6spy.multiline=true
